+++ poetry/console/commands/env/__init__.py
@@ -1,1 +1,2 @@
 from .info import EnvInfoCommand
+from .use import EnvUseCommand