+++ poetry/console/commands/env_command.py
@@ -9,13 +9,14 @@ class EnvCommand(Command):
 
     def initialize(self, i, o):
         from poetry.semver import parse_constraint
-        from poetry.utils.env import Env
+        from poetry.utils.env import EnvManager
 
         super(EnvCommand, self).initialize(i, o)
 
         # Checking compatibility of the current environment with
         # the python dependency specified in pyproject.toml
-        current_env = Env.get(self.poetry.file.parent)
+        env_manager = EnvManager(self.poetry.config)
+        current_env = env_manager.get(self.poetry.file.parent)
         supported_python = self.poetry.package.python_constraint
         current_python = parse_constraint(
             ".".join(str(v) for v in current_env.version_info[:3])
@@ -29,7 +30,7 @@ class EnvCommand(Command):
                 )
             )
 
-        self._env = Env.create_venv(
+        self._env = env_manager.create_venv(
             self.poetry.file.parent, o, self.poetry.package.name
         )